@@ -1,9 +1,11 @@
-﻿List<Racer> racers = new();
-racers.Add(new Racer(26, "Jacques", "Villeneuve", "Canada", 11));
-racers.Add(new Racer(18, "Alan", "Jones", "Australia", 12));
-racers.Add(new Racer(11, "Jackie", "Stewart", "United Kingdom", 27));
-racers.Add(new Racer(15, "James", "Hunt", "United Kingdom", 10));
-racers.Add(new Racer(5, "Jack", "Brabham", "Australia", 14));
+﻿List<Racer> racers =
+[
+    new Racer(26, "Jacques", "Villeneuve", "Canada", 11),
+    new Racer(18, "Alan", "Jones", "Australia", 12),
+    new Racer(11, "Jackie", "Stewart", "United Kingdom", 27),
+    new Racer(15, "James", "Hunt", "United Kingdom", 10),
+    new Racer(5, "Jack", "Brabham", "Australia", 14),
+];
 
 var lookupRacers = racers.ToLookup(r => r.Country);
 
