@@ -1,13 +1,7 @@
-﻿sealed class AliceRunner : IDisposable
+﻿sealed class AliceRunner(ILogger<AliceRunner> logger) : IDisposable
 {
-    private readonly ILogger _logger;
-    private readonly ECDiffieHellman _algorithm;
-    public AliceRunner(ILogger<AliceRunner> logger)
-    {
-        _logger = logger;
-        _algorithm = ECDiffieHellman.Create();
-        _logger.LogInformation($"Using this ECDiffieHellman class: {_algorithm.GetType().Name}");
-    }
+    private readonly ILogger _logger = logger;
+    private readonly ECDiffieHellman _algorithm = ECDiffieHellman.Create();
 
     public void Dispose() => _algorithm.Dispose();
 
@@ -15,6 +9,8 @@
 
     public async Task<(byte[] Iv, byte[] EncryptedData)> GetSecretMessageAsync(ECDiffieHellmanPublicKey otherPublicKey)
     {
+        _logger.LogInformation("Using this ECDiffieHellman class: {type}", _algorithm.GetType().Name);
+
         string message = "secret message from Alice";
         _logger.LogInformation("Alice sends message {message}", message);
 
@@ -25,7 +21,7 @@
             "Bobs public key information: {key}", Convert.ToBase64String(symmKey));
 
         using Aes aes = Aes.Create();
-        _logger.LogInformation("Using this Aes class: {class}", aes.GetType().Name);
+        _logger.LogInformation("Using this Aes class: {type}", aes.GetType().Name);
         aes.Key = symmKey;
         aes.GenerateIV();
         using ICryptoTransform encryptor = aes.CreateEncryptor();
