@@ -4,13 +4,15 @@ using Microsoft.Extensions.Hosting;
 using var host = Host.CreateDefaultBuilder(args)
     .ConfigureServices(services =>
     {
-        services.AddTransient<AliceRunner>();
-        services.AddTransient<BobRunner>();
+        services.AddScoped<AliceRunner>();
+        services.AddScoped<BobRunner>();
     })
     .Build();
 
-var alice = host.Services.GetRequiredService<AliceRunner>();
-var bob = host.Services.GetRequiredService<BobRunner>();
+using var scope1 = host.Services.CreateScope();
+
+var alice = scope1.ServiceProvider.GetRequiredService<AliceRunner>();
+var bob = scope1.ServiceProvider.GetRequiredService<BobRunner>();
 var keyAlice = alice.GetPublicKey();
 var aliceData = alice.GetDocumentAndSignature();
 bob.VerifySignature(aliceData.Data, aliceData.Sign, keyAlice);
